TER-78 separate common filters to base class
